First commit:
This is a first commit with a very basic unstyled index page. Bootstrap 4 is being imported in the head via the Bootstrap CDN. This needs to be changed so we are using a local copy of Bootstrap. This commit is really just to get an initial file in the repo that we can begin developing.