Sign up security began:
I have added a basic honeypot to the sign up form. I have also added a thank you page that will be redirected to when a user successfully signs up. This needs to be styled. There is also an issue with the php code for the sign up not executing properly.