Login working with Firefox:
The login is working in a basic fashion with firefox but the autofill feature in chrome is giving issues. I have also removed the email and password text as it was throwing off the styling and the placeholders are working in Firefox. Need to come up with a solution to turn off autofill in Chrome and everything will be working acceptably.