Added document sample funtionality
Users creating a task are now required to also provide a sample of the document. This sample is stored in a new column in  the document table and is displayed on the task-details page. If the document sample is a pdf we embed the document on the page otherwise we provide a link to download the sample document.